+++ geefcc/ee_tmf.py
@@ -77,5 +77,4 @@ def ee_tmf(years):
 
     return forest
 
-
 # End